# Backend/user-service/src/main/java/com/codechallenge/user/user_service/dto/SubmissionHistoryDTO.java
package com.codechallenge.user.user_service.dto;

public class SubmissionHistoryDTO {
    private Long id;
    private Long problemId;
    private String language;
    private String status;
    private Integer score;
    private String submittedAt;

    // Getters and setters
    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public Long getProblemId() { return problemId; }
    public void setProblemId(Long problemId) { this.problemId = problemId; }
    public String getLanguage() { return language; }
    public void setLanguage(String language) { this.language = language; }
    public String getStatus() { return status; }
    public void setStatus(String status) { this.status = status; }
    public Integer getScore() { return score; }
    public void setScore(Integer score) { this.score = score; }
    public String getSubmittedAt() { return submittedAt; }
    public void setSubmittedAt(String submittedAt) { this.submittedAt = submittedAt; }
} 