# Backend/user-service/src/main/java/com/codechallenge/user/user_service/dto/UserUpdateRequest.java
package com.codechallenge.user.user_service.dto;

public class UserUpdateRequest {
    private String username;
    private String avatar;

    // Getters and setters
    public String getUsername() { return username; }
    public void setUsername(String username) { this.username = username; }
    public String getAvatar() { return avatar; }
    public void setAvatar(String avatar) { this.avatar = avatar; }
} 